apply some feedback
