+++ CHANGELOG.md
@@ -9,12 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Centralize ad-hoc `http.Client` instantiations into a single shared `common.HTTPClient`, giving one source of truth for outbound client configuration.
 
 ### Added
-- Support for the IETF OAuth Token Status List (draft-ietf-oauth-status-list-15) on SD-JWT VC credentials
-  - New `eudi/credentials/statuslist` package: fetches, verifies, and decodes Status List Tokens (`application/statuslist+jwt`) over HTTP, with singleflight dedup, body-size caps, and zlib decoding for all four bit-sizes (1, 2, 4, 8)
-  - Status List Token signatures are validated via the same `x5c`-or-`kid`+DID (did:web/did:jwk) dispatcher as the SD-JWT VC itself, refactored into the shared `eudi/jwt.JwtKeyProvider`. The token is bound to the credential via its `sub` claim (which must equal the referenced `uri`) plus a trusted signature; a delegated Status Issuer signing with its own key is supported, per draft-15 §11.3
-  - Status checks run at every credential lifecycle site: holder-side at issuance (fail-closed), verifier-side after SD-JWT VC verification, and a background sweep (`Client.RefreshStatuses`) keeps stored credentials' `LastKnownStatus` up to date. The background sweep checks one representative instance per batch — a batch's copies are the same credential and are revoked together, so one status entry stands in for the whole batch. At disclosure the wallet does not fail closed — a revoked instance is surfaced on the OpenID4VP disclosure plan with `Revoked=true` (matching IRMA), leaving the decision to the frontend and the verifier's own status check
-  - Persistent SQLCipher-backed cache (`status_list_cache` table) lets the wallet survive cold starts and brief offline periods within a list's TTL. TTL comes from the token's own `ttl`/`exp` (draft-15 §8.2), falling back to the HTTP `max-age` when the token advertises neither, clamped to `[60s, 24h]`. Fail-closed on fetch/verify/decode errors past TTL
-  - The `Client` exposes a `RefreshStatuses(ctx)` method for UI-initiated refreshes
+- Support for the IETF OAuth Token Status List (draft-ietf-oauth-status-list-15) on SD-JWT VC credentials: the wallet fetches, verifies, and caches Status List Tokens (`application/statuslist+jwt`), checks credential status at issuance, at disclosure, and on a background sweep, and exposes `Client.RefreshStatuses(ctx)` for UI-initiated refreshes.
 - `eudi/holderkeys`: a CGO-free package providing the holder-key seam (`HolderSigner`, `SoftwareHolderSigner`, the KB-JWT `NewSignerKeyBinder` bridge) so a WSCA adapter or a server-side (Postgres) holder can implement external holder-key signing without pulling in a sqlcipher (cgo) dependency.
 - Pluggable holder-key binding seams for external secure devices (WSCA/HSM): `openid4vci.NewClient` takes a required `HolderKeyBinder` and `eudi_sdjwt_dcql.NewSdJwtVcDcqlHandler` a required `sdjwtvc.KeyBinder`, and `proofs.BuildWithES256Signer` signs the OpenID4VCI proof of possession via an external signer. Callers pass the software, storage-backed binder for the existing behaviour, or a WSCA/HSM-backed implementation.
 - `storage.NewStorageWithDialector(dialector, fs)`: open the EUDI holder database on any GORM dialector (e.g. `gorm.io/driver/postgres`) rather than only sqlcipher, for server-side / multi-tenant deployments. `NewStorage` is unchanged (it builds the sqlcipher dialector and delegates). The caller owns the at-rest encryption posture of a non-sqlcipher driver.
